created spoiler for account balance when shake
