+++ lib/data/database/database.dart
@@ -2,7 +2,6 @@ import 'dart:io';
 
 import 'package:drift/drift.dart';
 import 'package:drift/native.dart';
-import 'package:drift_flutter/drift_flutter.dart';
 import 'package:path/path.dart' as p;
 import 'package:path_provider/path_provider.dart';
 
@@ -28,11 +27,5 @@ class AppDatabase extends _$AppDatabase {
       final file = File(p.join(dbFolder.path, 'db.sqlite'));
       return NativeDatabase(file);
     });
-    return driftDatabase(
-      name: 'accountPage_database',
-      native: const DriftNativeOptions(
-        databaseDirectory: getApplicationSupportDirectory,
-      ),
-    );
   }
 }
