+++ lib/presentation/pages/bank_account_page/bank_account_page.dart
@@ -6,8 +6,12 @@ import 'package:go_router/go_router.dart';
 import 'package:provider/provider.dart';
 import 'package:ya_finance_app/data/database/account_db.dart';
 import 'package:ya_finance_app/data/repositories_impl/bank_account_mock.dart';
+import 'package:ya_finance_app/presentation/widgets/spoiler/spoiler.dart';
+import '../../widgets/currency_picker/currency_picker.dart';
 import 'account_provider.dart';
 import 'bloc/ba_page_bloc.dart';
+import 'package:shake/shake.dart';
+
 
 class BankAccountPage extends StatefulWidget {
   const BankAccountPage({super.key});
@@ -18,11 +22,28 @@ class BankAccountPage extends StatefulWidget {
 
 class _BankAccountPageState extends State<BankAccountPage> {
   final _baPageBloc = BaPageBloc(MockBankAccountRepository(), DBAccountRepository());
+  bool _isTextVisible = true;
+  late ShakeDetector detector;
 
   @override
   void initState() {
     _baPageBloc.add(LoadAccountEvent());
+    detector = ShakeDetector.autoStart(
+      onPhoneShake: (ShakeEvent event) {
+        setState(() {
+          _isTextVisible = !_isTextVisible;
+        });
+      },
+      minimumShakeCount: 2,
+      shakeSlopTimeMS: 350,
+    );
     super.initState();
+  }
+
+  @override
+  void dispose() {
+    detector.stopListening();
+    super.dispose();
   }
 
   @override
@@ -103,13 +124,13 @@ class _BankAccountPageState extends State<BankAccountPage> {
                           ),
                           Row(
                             children: [
-                              Text(
+                              (_isTextVisible) ? Text(
                                 '${Provider.of<AccountProvider>(context).balance} ${Provider.of<AccountProvider>(context).currency}',
                                 style: TextStyle(
                                   fontSize: 20,
                                   fontWeight: FontWeight.w400,
                                 ),
-                              ),
+                              ) : ParticleContainer(width: 100, height: 20,),
                               Container(width: 20),
                               Icon(
                                 Icons.navigate_next,
@@ -142,7 +163,7 @@ class _BankAccountPageState extends State<BankAccountPage> {
                         ),
                       ),
                       onPressed: () async {
-                        final currency = await _showCurrencyPicker(context);
+                        final currency = await showCurrencyPicker(context);
                         if (currency != null) {
                           Provider.of<AccountProvider>(context, listen: false).changeCurrency(currency);
                         }
@@ -215,195 +236,4 @@ class _BankAccountPageState extends State<BankAccountPage> {
       ),
     );
   }
-}
-
-Future<String?> _showCurrencyPicker(BuildContext context) async {
-  String? cur = null;
-  final dialog = await showModalBottomSheet<String?>(
-    context: context,
-    builder: (BuildContext context) {
-      return Container(
-        height: MediaQuery.of(context).size.height * 0.35,
-        decoration: BoxDecoration(
-          color: Color(0xFFF7F2FA),
-          borderRadius: BorderRadius.vertical(top: Radius.circular(16)),
-        ),
-        child: Column(
-          children: [
-            Padding(
-              padding: EdgeInsets.symmetric(vertical: 20),
-              child: Container(
-                width: 45,
-                height: 5,
-                decoration: BoxDecoration(
-                  color: Color(0xFF79747E),
-                  borderRadius: BorderRadius.circular(16),
-                ),
-              ),
-            ),
-            Expanded(
-              flex: 1,
-              child: Container(
-                decoration: BoxDecoration(
-                  border: Border(
-                    bottom: BorderSide(color: Color(0xFFCAC4D0), width: 1),
-                  ),
-                ),
-                child: ElevatedButton(
-                  style: ElevatedButton.styleFrom(
-                    backgroundColor: Color(0xFFF7F2FA),
-                    elevation: 0,
-                    padding: EdgeInsets.symmetric(horizontal: 16, vertical: 20),
-                    shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(0),
-                    ),
-                  ),
-                  onPressed: () {
-                    cur = '₽';
-                    Navigator.pop(context);
-                  },
-                  child: Row(
-                    children: [
-                      Text(
-                        '₽',
-                        style: TextStyle(
-                          fontSize: 30,
-                          fontWeight: FontWeight.w400,
-                          color: Colors.black,
-                        ),
-                      ),
-                      SizedBox(width: 20),
-                      Text(
-                        'Российский рубль ₽',
-                        style: TextStyle(
-                          fontSize: 20,
-                          fontWeight: FontWeight.w400,
-                          color: Colors.black,
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            ),
-            Expanded(
-              flex: 1,
-              child: Container(
-                decoration: BoxDecoration(
-                  border: Border(
-                    bottom: BorderSide(color: Color(0xFFCAC4D0), width: 1),
-                  ),
-                ),
-                child: ElevatedButton(
-                  style: ElevatedButton.styleFrom(
-                    backgroundColor: Color(0xFFF7F2FA),
-                    elevation: 0,
-                    padding: EdgeInsets.symmetric(horizontal: 16, vertical: 20),
-                    shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(0),
-                    ),
-                  ),
-                  onPressed: () {
-                    cur = '\$';
-                    Navigator.pop(context);
-                  },
-                  child: Row(
-                    children: [
-                      Text(
-                        '\$',
-                        style: TextStyle(
-                          fontSize: 30,
-                          fontWeight: FontWeight.w400,
-                          color: Colors.black,
-                        ),
-                      ),
-                      SizedBox(width: 20),
-                      Text(
-                        'Американский доллар \$',
-                        style: TextStyle(
-                          fontSize: 20,
-                          fontWeight: FontWeight.w400,
-                          color: Colors.black,
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            ),
-            Expanded(
-              flex: 1,
-              child: ElevatedButton(
-                style: ElevatedButton.styleFrom(
-                  backgroundColor: Color(0xFFF7F2FA),
-                  elevation: 0,
-                  padding: EdgeInsets.symmetric(horizontal: 16, vertical: 20),
-                  shape: RoundedRectangleBorder(
-                    borderRadius: BorderRadius.circular(0),
-                  ),
-                ),
-                onPressed: () {
-                  cur = '€';
-                  Navigator.pop(context);
-                },
-                child: Row(
-                  children: [
-                    Text(
-                      '€',
-                      style: TextStyle(
-                        fontSize: 30,
-                        fontWeight: FontWeight.w400,
-                        color: Colors.black,
-                      ),
-                    ),
-                    SizedBox(width: 20),
-                    Text(
-                      'Евро',
-                      style: TextStyle(
-                        fontSize: 20,
-                        fontWeight: FontWeight.w400,
-                        color: Colors.black,
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-            Expanded(
-              flex: 1,
-              child: ElevatedButton(
-                style: ElevatedButton.styleFrom(
-                  backgroundColor: Color(0xFFE46962),
-                  elevation: 0,
-                  padding: EdgeInsets.symmetric(horizontal: 16, vertical: 20),
-                  shape: RoundedRectangleBorder(
-                    borderRadius: BorderRadius.circular(0),
-                  ),
-                ),
-                onPressed: () {
-                  cur = null;
-                  Navigator.pop(context);
-                },
-                child: Row(
-                  children: [
-                    Icon(Icons.cancel_outlined, color: Colors.white, size: 30),
-                    SizedBox(width: 10),
-                    Text(
-                      'Отмена',
-                      style: TextStyle(
-                        fontSize: 20,
-                        fontWeight: FontWeight.w400,
-                        color: Colors.white,
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-          ],
-        ),
-      );
-    },
-  );
-  return cur;
 }
