+++ lib/presentation/pages/bank_account_page/bloc/ba_page_bloc.dart
@@ -1,9 +1,7 @@
 import 'dart:async';
 import 'dart:io';
 import 'package:bloc/bloc.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 import 'package:ya_finance_app/data/repositories_impl/bank_account_mock.dart';
-import 'package:connectivity_plus/connectivity_plus.dart';
 
 import '../../../../data/database/account_db.dart';
