+++ pubspec.yaml
@@ -29,7 +29,6 @@ dependencies:
   path_provider: ^2.1.5
   sqlite3_flutter_libs: ^0.5.34
   shared_preferences: ^2.5.3
-  connectivity_plus: ^4.0.2
   shake: ^3.0.0
